+++ src/control/HdrHexLogoImg/template.mjs
@@ -1,65 +1,19 @@
 import ControlMaker from '../../lib/ControlMaker.mjs';
+import { HdrBaseLogoImg } from '../../comp/HdrBaseLogoImg.mjs';
 
-import { DARKMODE_SELECTOR_VALUE } from '../../lib/DarkMode.mjs';
-import { HEADER_MOBILE_DEVICE_WIDTH } from '../../lib/WickedTheme.mjs';
+const NAME = "HdrHexLogoImg";
 
-const mk = new ControlMaker('HdrHexLogoImg', import.meta.url);
-
-const width = 225;
-const height = 225;
-
-const clss = mk.newClassNameMap([
-  "ROOT_CLASS",
-]);
-
-const vars = mk.newCSSVariableMap({
-  favicon: [
-    await mk.loadSvgAsCssUrl('./favicon1.svg'),
-    await mk.loadSvgAsCssUrl('./favicon2.svg'),
-  ],
-});
-
-mk.newHTML('ROOT_HTML', `
-  <h3 class="${clss.ROOT_CLASS}"></h3>
-`);
-
-mk.newHTML('CSS', `
-:root
+export async function buildComponent()
 {
-  ${vars.toString(0)};
-}
-
-${DARKMODE_SELECTOR_VALUE}
-{
-  ${vars.toString(1)};
-}
-
-.${clss.ROOT_CLASS}
-{
-  height: 100%;
-  width: 56px;
-  margin: 0 7px 0 0;
-  background-size: contain;
-  background-repeat: no-repeat;
-  background-position-y: center;
-  background-position-x: center;
-  background-image: ${vars.favicon.asVar()};
-  flex-shrink: 0;
-  padding: 0;
-  font-size: 1em;
-  font-weight: 400;
-}
-
-@media (device-width < ${HEADER_MOBILE_DEVICE_WIDTH})
-{
-  .${clss.ROOT_CLASS}
-  {
-    display: none;
-  }
-}
-`);
-
-export function buildComponent()
-{
+  const mk = new ControlMaker(NAME, import.meta.url);
+  const comp = new HdrBaseLogoImg({
+    favicon: [
+      await mk.loadSvgAsCssUrl('./favicon1.svg'),
+      await mk.loadSvgAsCssUrl('./favicon2.svg'),
+    ],
+    width: 56,
+  });
+  
+  comp.make(mk);
   return mk.buildComponent();
 }
